old-tests/submission/W3C/domhtml/HTMLScriptElement01.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>script</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/the-script-element.html#dom-script-text" rel="help">
<link rel="author" title="W3C">
</head>
<body>
<p>The IDL attribute text must return a concatenation of the contents of all the Text nodes that are children of the script element (ignoring any other nodes such as comments or elements), in tree order. On setting, it must act the same way as the textContent IDL attribute.</p>

<script>var a=2;</script>

<script>
test(function () {
  var nodeList;
  var testNode;
  var vtext;
  var doc;

      doc = window.document;
      nodeList = doc.getElementsByTagName("script");
      assert_equals(nodeList.length,4);
      testNode = nodeList.item(2);
      assert_equals(testNode.text,"var a=2;");
      assert_equals(a,2);
      testNode.text="var a=3;";
      assert_equals(testNode.textContent,"var a=3;");
      assert_equals(a,2);

}, "The text attribute specifies the script content of the element.");
</script>
<div id="log"></div>
</body></html>
